# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resource Booking API",
        "description": "API documentation for Resource Booking System",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Создание бронирования",
                "description": "Создает новое бронирование для ресурса",
                "operationId": "createBooking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resource_id",
                                    "user_id",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "resource_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-02-17T10:00:00Z"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-02-17T11:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Бронирование создано успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/bookings/{id}": {
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Отмена бронирования",
                "description": "Отменяет бронирование по его ID",
                "operationId": "deleteBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID бронирования",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Бронирование успешно отменено"
                    },
                    "404": {
                        "description": "Бронирование не найдено"
                    }
                }
            }
        },
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Получение списка ресурсов",
                "description": "Возвращает список всех ресурсов",
                "operationId": "getResourcesList",
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Resource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Создание ресурса",
                "description": "Создает новый ресурс",
                "operationId": "createResource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Conference Room"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "room"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание конференц-зала"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ресурс создан успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/resources/{id}/bookings": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Получение бронирований для ресурса",
                "description": "Возвращает список бронирований для указанного ресурса",
                "operationId": "getResourceBookings",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ресурса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Booking"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ресурс не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Booking": {
                "title": "Booking",
                "required": [
                    "id",
                    "resource_id",
                    "user_id",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "id": {
                        "description": "Class Booking\n\nМодель, представляющая бронирование ресурса.",
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "resource_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-17T10:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-17T11:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-02-17T09:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-02-17T09:00:00Z"
                    }
                },
                "type": "object"
            },
            "Resource": {
                "title": "Resource",
                "required": [
                    "id",
                    "name",
                    "type"
                ],
                "properties": {
                    "id": {
                        "description": "Class Resource\n\nМодель, представляющая ресурс, который можно бронировать.",
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Conference Room"
                    },
                    "type": {
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание конференц-зала",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-02-17T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-02-17T10:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Bookings",
            "description": "Bookings"
        },
        {
            "name": "Resources",
            "description": "Resources"
        }
    ]
}